// lib/enums/http_methods.ts
/**
 * Enum containing the
 * available HTTP methods.
 * 
 * @since 19/03/2023
 * @author Felipe Matheus Flohr
 */
enum HttpMethod {
    GET = "GET",
    HEAD = "HEAD",
    POST = "POST",
    PUT = "PUT",
    DELETE = "DELETE",
    CONNECT = "CONNECT",
    OPTIONS = "OPTIONS",
    TRACE = "TRACE",
    PATCH = "PATCH"
}

export default HttpMethod;